+++ webpack.config.js
@@ -1,0 +1,100 @@
+const webpack = require('webpack');
+const path = require('path');
+const fileSystem = require('fs');
+const env = require('./utils/env');
+const CleanWebpackPlugin = require('clean-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WriteFilePlugin = require('write-file-webpack-plugin');
+const CopyPlugin = require('copy-webpack-plugin');
+
+// load the secrets
+const alias = {};
+const secretsPath = path.join(__dirname, ('secrets.' + env.NODE_ENV + '.js'));
+const fileExtensions = ['jpg', 'jpeg', 'png', 'gif', 'eot', 'otf', 'svg', 'ttf', 'woff', 'woff2'];
+
+if (fileSystem.existsSync(secretsPath)) {
+  alias['secrets'] = secretsPath;
+}
+
+var options = {
+  mode: process.env.NODE_ENV || 'development',
+  entry: {
+    popup: path.join(__dirname, 'src', 'js', 'popup', 'popup.js'),
+    options: path.join(__dirname, 'src', 'js', 'options.js'),
+    background: path.join(__dirname, 'src', 'js', 'background.js'),
+    content_script: path.join(__dirname, 'src', 'js', 'content_script', 'content.js')
+  },
+  output: {
+    path: path.join(__dirname, 'build'),
+    filename: '[name].bundle.js'
+  },
+  chromeExtensionBoilerplate: {
+    notHotReload: ['content_script']
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        loader: 'style-loader!css-loader',
+        exclude: /node_modules/
+      },
+      {
+        test: /\.less$/,
+        use: [
+          'style-loader',
+          'css-loader',
+          'less-loader'
+        ],
+      },
+      {
+        test: new RegExp('\.(' + fileExtensions.join('|') + ')$'),
+        loader: 'file-loader?name=[name].[ext]',
+        exclude: /node_modules/
+      },
+      {
+        test: /\.html$/,
+        loader: 'html-loader',
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    alias: alias
+  },
+  plugins: [
+    // clean the build folder
+    new CleanWebpackPlugin(['build']),
+    // expose and write the allowed env vars on the compiled bundle
+    new webpack.EnvironmentPlugin(['NODE_ENV']),
+    new CopyPlugin([
+      {
+        from: 'src/manifest.json'
+      },
+      {
+        from: 'src/img/'
+      }
+    ]),
+    new HtmlWebpackPlugin({
+      template: path.join(__dirname, 'src', 'popup.html'),
+      filename: 'popup.html',
+      chunks: ['popup']
+    }),
+    new HtmlWebpackPlugin({
+      template: path.join(__dirname, 'src', 'options.html'),
+      filename: 'options.html',
+      chunks: ['options']
+    }),
+    new HtmlWebpackPlugin({
+      template: path.join(__dirname, 'src', 'background.html'),
+      filename: 'background.html',
+      chunks: ['background']
+    }),
+    new WriteFilePlugin()
+  ]
+};
+
+if (env.NODE_ENV === 'development') {
+  options.devtool = 'cheap-module-eval-source-map';
+}
+
+module.exports = options;